requirements.txt
intasend-python
flask
flask-cors
python-dotenv
requests
